Finished Porting
All code from the previous C code base has been ported to AGK format Basic. It compiles, but does not run. This appears to be due to the fact that the C code base had extremely tight RAM requirements. I will now work through expanding the code to avoid using all the logical and bitwise operators to conserve RAM, and will instead focus on making it work as expected.